Add AndroidApplicationConfigurator
